feat: fix type declaration
