+++ examples/node/cjs/index.js
@@ -1,4 +1,4 @@
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const {cep, Provider, factory, CepService} = require('../../../dist/cjs/index')
+const { cep, Provider, factory, CepService, Requester } = require('../../../dist/cjs/index')
 cep('41342315').then(console.log)
-console.log(Provider, factory, CepService)
+console.log(cep, Provider, factory, CepService, Requester)
