+++ src/cep.ts
@@ -1,6 +1,7 @@
+import { Cep } from './entity/'
 import factory from './factory'
 
-export const cep = (cep: string) => {
+export const cep = (cep: string): Promise<Cep> => {
 	const handler = factory({
 		useDefaultProviders: true,
 	})
