+++ src/entity/cep.ts
@@ -1,0 +1,7 @@
+export type Cep = {
+	cep: string,
+	street: string,
+	state: string,
+	neighborhood: string
+	city: string
+}
