+++ src/entity/index.ts
@@ -1,7 +1,3 @@
-export type Cep = {
-	cep: string,
-	street: string,
-	state: string,
-	neighborhood: string
-	city: string
-}
+import { Cep } from './cep'
+
+export { Cep }
