--- src/lib.ts
@@ -1,13 +1,0 @@
-import { CepService } from './service/index'
-import { cep } from './cep'
-import factory from './factory'
-import { Provider } from './provider'
-import { Requester } from './requester/index'
-
-export default {
-	cep,
-	CepService,
-	Provider,
-	factory,
-	Requester,
-}
